backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
python-decouple==3.8
Pillow==12.0.0





